+++ src/test/java/com/contextualmocker/core/ClassWithFinalMethod.java
@@ -1,0 +1,11 @@
+package com.contextualmocker.core;
+
+public class ClassWithFinalMethod {
+    public String sayHello() {
+        return "Hello";
+    }
+
+    public final String finalMethod() {
+        return "This is final";
+    }
+}