+++ src/test/java/com/contextualmocker/core/ClassWithStaticMethod.java
@@ -1,0 +1,11 @@
+package com.contextualmocker.core;
+
+public class ClassWithStaticMethod {
+    public String sayHello() {
+        return "Hello";
+    }
+
+    public static String staticMethod() {
+        return "Static";
+    }
+}