+++ src/test/java/com/contextualmocker/core/FinalClass.java
@@ -1,0 +1,7 @@
+package com.contextualmocker.core;
+
+public final class FinalClass {
+    public String sayHello() {
+        return "Hello from final class";
+    }
+}